Add an image, a description and lists
I added a linked image of a Mac and Cheese and a description of
the plate that I have found on allrecipes.com. I also added an
unordered list of the ingredients we need to make the plate and
an ordered list of the steps to follow for making it. Both have
also been found on allrecipes.com.